Reorganize assets location
