+++ laravel/resources/css/app.css
@@ -1,3 +1,5 @@
+@import url('./editor.css');
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
