+++ laravel/resources/js/app.js
@@ -1,2 +1,1 @@
 import './bootstrap';
-import './editor';
