+++ laravel/resources/js/editor.js
@@ -1,273 +1,364 @@
-import { EditorState, Compartment, StateField, StateEffect } from "@codemirror/state";
-import { EditorView, lineNumbers, drawSelection, highlightActiveLine, Decoration, WidgetType } from "@codemirror/view";
-import { foldGutter } from "@codemirror/language";
-import { xml } from "@codemirror/lang-xml";
-import { oneDark } from "@codemirror/theme-one-dark";
+import initEditor from './codemirror-editor';
 
-// Widget to replace tags with invisible content
-class InvisibleTagWidget extends WidgetType {
-  constructor(tag) {
-    super();
-    this.tag = tag;
-  }
+document.addEventListener('DOMContentLoaded', () => {
+    const { xmlContent, comparisonId, fileType, canEdit } = window.editorParams;
 
-  toDOM() {
-    const span = document.createElement("span");
-    span.style.display = "none";
-    span.textContent = this.tag;
-    return span;
-  }
-
-  ignoreEvent() {
-    return false;
-  }
-}
-
-// Effect to toggle tag visibility
-const toggleTagVisibility = StateEffect.define();
-
-// State field to track whether tags should be hidden
-const hideTagsField = StateField.define({
-  create() {
-    return true; // Start with tags hidden
-  },
-  update(value, tr) {
-    for (let effect of tr.effects) {
-      if (effect.is(toggleTagVisibility)) {
-        value = effect.value;
-      }
-    }
-    return value;
-  },
-  provide: f => EditorView.decorations.from(f, hideTags => {
-    if (!hideTags) return Decoration.none;
-
-    return view => {
-      const widgets = [];
-      const doc = view.state.doc;
-      const text = doc.toString();
-
-      // Regex to match XML tags: <tag...> or </tag> or <tag/>
-      const tagRegex = /<\/?[a-zA-Z][^>]*\/?>/g;
-      let match;
-
-      while ((match = tagRegex.exec(text)) !== null) {
-        const from = match.index;
-        const to = from + match[0].length;
-
-        widgets.push(
-          Decoration.replace({
-            widget: new InvisibleTagWidget(match[0]),
-            inclusive: true,
-            block: false
-          }).range(from, to)
-        );
-      }
-
-      return Decoration.set(widgets);
+    // DOM Elements
+    const elements = {
+        saveBtn: document.getElementById('save-xml'),
+        toggleBtn: document.getElementById('toggle-readonly'),
+        toggleTagsBtn: document.getElementById('toggle-tags'),
+        previewImg: document.getElementById('facsimile-preview'),
+        noPreviewText: document.getElementById('no-preview'),
+        loadingSpinner: document.getElementById('loading-spinner'),
+        editorContainer: document.getElementById('editor-container'),
+        imageName: document.getElementById('image-name'),
+        pagination: document.getElementById('pagination'),
     };
-  })
-});
 
-// Helper function to create page number decorations
-function createPageNumberDecorations(state) {
-  const decorations = [];
-  const text = state.doc.toString();
+    // Constants
+    const MESSAGES = {
+        DEFAULT: 'Cliquez sur un bouton "Insérer..." pour activer l\'insertion',
+        ERROR: 'Erreur lors du chargement de l\'image'
+    };
 
-  // Regex to match page number content: <span class="page-number">XXX</span>
-  const pageNumberRegex = /<span class="page-number">([^<]*)<\/span>/g;
-  let match;
+    const BUTTON_STATES = {
+        INACTIVE: { add: 'btn-primary', remove: ['btn-warning', 'btn-success', 'btn-danger'] },
+        ACTIVE_INSERT: { add: 'btn-warning', remove: ['btn-primary', 'btn-success', 'btn-danger'] },
+        INSERTED: { add: 'btn-success', remove: ['btn-primary', 'btn-warning', 'btn-danger'] },
+        ACTIVE_DELETE: { add: 'btn-danger', remove: ['btn-primary', 'btn-warning', 'btn-success'] }
+    };
 
-  while ((match = pageNumberRegex.exec(text)) !== null) {
-    const contentStart = match.index + '<span class="page-number">'.length;
-    const contentEnd = contentStart + match[1].length;
+    // State - only one active button at a time, either in insert or delete mode
+    let activeButton = null;
+    let isDeleteMode = false; // true = delete mode, false = insert mode
+    let lastImageSrc = null; // Track the last image displayed
+    const itemsPerPage = 3;
 
-    decorations.push(
-      Decoration.mark({
-        class: "cm-page-number"
-      }).range(contentStart, contentEnd)
-    );
-  }
+    const editor = initEditor(document.getElementById('editor-container'), xmlContent);
 
-  return Decoration.set(decorations);
-}
-
-// State field to highlight page numbers
-const pageNumberHighlight = StateField.define({
-  create(state) {
-    return createPageNumberDecorations(state);
-  },
-  update(decorations, tr) {
-    // Only recalculate if the document has changed
-    if (tr.docChanged) {
-      return createPageNumberDecorations(tr.state);
+    if (!canEdit) {
+        editor.setReadOnly(true);
     }
-    return decorations;
-  },
-  provide: f => EditorView.decorations.from(f)
-});
 
-window.initEditor = (initialXml, comparisonId, fileType = 'source') => {
-  const container = document.getElementById('editor-container');
-  const saveBtn = document.getElementById('save-xml');
-
-  // Create compartments for dynamic reconfiguration
-  const readOnlyCompartment = new Compartment();
-  const editableCompartment = new Compartment();
-
-  const startState = EditorState.create({
-    doc: initialXml,
-    extensions: [
-      xml(),
-      lineNumbers(),
-      foldGutter(),
-      oneDark,
-      EditorView.lineWrapping,
-      drawSelection(),
-      hideTagsField, // Add the tag hiding field
-      pageNumberHighlight, // Add the page number highlighting field
-      readOnlyCompartment.of(EditorState.readOnly.of(true)),
-      editableCompartment.of(EditorView.editable.of(false)),
-      EditorView.theme({
-        ".cm-cursor": {
-          borderLeftColor: "#528bff !important",
-          borderLeftWidth: "2px !important",
-          display: "block !important",
-          visibility: "visible !important"
-        },
-        ".cm-selectionBackground": {
-          backgroundColor: "#2e4862ff !important"
-        },
-        ".cm-page-number": {
-          backgroundColor: "#ff6b35 !important",
-          color: "#ffffff !important",
-          fontWeight: "bold !important",
-          padding: "2px 6px !important",
-          borderRadius: "3px !important",
-          border: "2px solid #ff4500 !important",
-          display: "inline-block !important",
-          fontSize: "1.1em !important"
-        },
-      }),
-    ]
-  });
-
-  const view = new EditorView({ state: startState, parent: container });
-
-  let isReadOnly = true;
-
-  window.editor = {
-    toggleReadOnly() {
-      isReadOnly = !isReadOnly;
-      view.dispatch({
-        effects: [
-          readOnlyCompartment.reconfigure(EditorState.readOnly.of(isReadOnly)),
-          editableCompartment.reconfigure(EditorView.editable.of(!isReadOnly))
-        ]
-      });
-      view.focus();
-      return isReadOnly;
-    },
-
-    setReadOnly(value) {
-      isReadOnly = value;
-      view.dispatch({
-        effects: [
-          readOnlyCompartment.reconfigure(EditorState.readOnly.of(isReadOnly)),
-          editableCompartment.reconfigure(EditorView.editable.of(!isReadOnly))
-        ]
-      });
-    },
-
-    toggleTagVisibility() {
-      const currentState = view.state.field(hideTagsField);
-      view.dispatch({
-        effects: toggleTagVisibility.of(!currentState)
-      });
-      return !currentState;
-    },
-
-    insertPageMarker(imageName, pageNumber = '001') {
-      const { head } = view.state.selection.main;
-
-      // Build the page marker tag
-      const pageMarkerTag = `<span class="page-marker" data-image-name="${imageName}"><span class="page-number">${pageNumber}</span><img src="/img/settings/page_right.svg" /></span>`;
-
-      // Insert at cursor position
-      view.dispatch({
-        changes: { from: head, insert: pageMarkerTag },
-        selection: { anchor: head + pageMarkerTag.length }
-      });
-      view.focus();
-    },
-
-    getPageMarkerTag(imageName) {
-      const content = view.state.doc.toString();
-      const regex = new RegExp(
-        `<span class="page-marker" data-image-name="${imageName}"><span class="page-number">.*?</span><img[^>]*></span>`,
-        'i'
-      );
-      const match = content.match(regex);
-      return match ? { tag: match[0], pos: content.indexOf(match[0]) } : null;
-    },
-
-    scrollToPageMarker(imageName) {
-      const result = this.getPageMarkerTag(imageName);
-      if (result) {
-        view.dispatch({
-          selection: { anchor: result.pos, head: result.pos + result.tag.length },
-          effects: EditorView.scrollIntoView(result.pos, { y: "center" })
+    // Pagination functions
+    const checkIfPageFullyInserted = (pageNumber) => {
+        const pageButtons = document.querySelectorAll(`.button-item[data-page="${pageNumber}"] button[data-tag]`);
+        if (pageButtons.length === 0) return false;
+        
+        let allInserted = true;
+        pageButtons.forEach(button => {
+            const imageName = button.getAttribute('data-tag');
+            if (!editor.isPageMarkerInserted(imageName)) {
+                allInserted = false;
+            }
         });
-        view.focus();
-      }
-    },
+        
+        return allInserted;
+    };
 
-    isPageMarkerInserted(imageName) {
-      return this.getPageMarkerTag(imageName) !== null;
-    },
-
-    countPageMarkerOccurrences(imageName) {
-      const content = view.state.doc.toString();
-      const regex = new RegExp(`data-image-name="${imageName}"`, 'g');
-      const matches = content.match(regex);
-      return matches ? matches.length : 0;
-    },
-
-    removePageMarker(imageName) {
-      const result = this.getPageMarkerTag(imageName);
-
-      if (result) {
-        const tagPos = result.pos;
-        const tagEnd = tagPos + result.tag.length;
-
-        // Remove immediately
-        view.dispatch({
-          changes: { from: tagPos, to: tagEnd, insert: '' }
+    const updatePaginationColors = () => {
+        document.querySelectorAll('#pagination .page-item').forEach((item, index) => {
+            const pageNumber = index + 1;
+            const link = item.querySelector('.page-link');
+            const isFullyInserted = checkIfPageFullyInserted(pageNumber);
+            
+            if (isFullyInserted) {
+                link.classList.add('page-fully-inserted');
+            } else {
+                link.classList.remove('page-fully-inserted');
+            }
         });
-      }
-    },
-  };
+    };
 
-  saveBtn.addEventListener('click', async () => {
-    const updatedXml = view.state.doc.toString();
-    const response = await fetch(`/comparison/${comparisonId}/editor?type=${fileType}`, {
-      method: 'PUT',
-      headers: {
-        'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content'),
-        'Content-Type': 'application/xml',
-        'Accept': 'application/json'
-      },
-      body: updatedXml
+    const initPagination = () => {
+        const allButtons = document.querySelectorAll('.button-item');
+        const totalPages = Math.ceil(allButtons.length / itemsPerPage);
+        
+        allButtons.forEach((item, index) => {
+            const itemPage = Math.ceil((index + 1) / itemsPerPage);
+            item.setAttribute('data-page', itemPage);
+        });
+        
+        if (totalPages <= 1) {
+            elements.pagination.parentElement.style.display = 'none';
+            showPage(1);
+            return;
+        }
+        
+        elements.pagination.innerHTML = '';
+        
+        for (let i = 1; i <= totalPages; i++) {
+            const li = document.createElement('li');
+            li.className = 'page-item';
+            
+            const link = document.createElement('a');
+            link.className = 'page-link shadow-none';
+            link.href = '#';
+            link.textContent = i;
+            link.addEventListener('click', (e) => {
+                e.preventDefault();
+                showPage(i);
+            });
+            
+            li.appendChild(link);
+            elements.pagination.appendChild(li);
+        }
+        
+        showPage(1);
+    };
+
+    const showPage = (page) => {
+        const allButtons = document.querySelectorAll('.button-item');
+        allButtons.forEach((item) => {
+            const itemPage = parseInt(item.getAttribute('data-page'));
+            item.style.display = itemPage === page ? '' : 'none';
+        });
+        
+        // Update pagination active state
+        document.querySelectorAll('#pagination .page-item').forEach((item, index) => {
+            if (index + 1 === page) {
+                item.classList.add('active');
+            } else {
+                item.classList.remove('active');
+            }
+        });
+    };
+
+    // Utility functions
+    const setButtonState = (button, state) => {
+        if (Array.isArray(state.remove)) {
+            button.classList.remove(...state.remove);
+        } else {
+            button.classList.remove(state.remove);
+        }
+        button.classList.add(state.add);
+    };
+
+    const showMessage = (message) => {
+        if (elements.noPreviewText) {
+            elements.noPreviewText.textContent = message;
+            elements.noPreviewText.style.display = 'block';
+        }
+    };
+    showMessage(MESSAGES.DEFAULT);
+
+    const deactivateActiveButton = () => {
+        if (activeButton) {
+            const imageName = activeButton.getAttribute('data-tag');
+            const isInserted = editor.isPageMarkerInserted(imageName);
+            
+            // Reset to appropriate state
+            if (isInserted) {
+                setButtonState(activeButton, BUTTON_STATES.INSERTED);
+            } else {
+                setButtonState(activeButton, BUTTON_STATES.INACTIVE);
+            }
+            activeButton.querySelector('span').textContent = imageName;
+            
+            activeButton = null;
+            isDeleteMode = false;
+            elements.editorContainer.style.cursor = 'default';
+        }
+    };
+
+    const loadImage = (imgSrc) => {
+        if (!imgSrc) return;
+        
+        lastImageSrc = imgSrc;
+        if (elements.noPreviewText) elements.noPreviewText.style.display = 'none';
+        
+        // Extract filename from URL
+        const filename = imgSrc.split('/').pop();
+        
+        // Set a timeout to show spinner only if loading takes more than 500ms
+        let spinnerTimeout = setTimeout(() => {
+            elements.previewImg.style.display = 'none';
+            elements.loadingSpinner.style.display = 'block';
+        }, 500);
+        
+        const img = new Image();
+        img.onload = () => {
+            clearTimeout(spinnerTimeout);
+            elements.loadingSpinner.style.display = 'none';
+            elements.previewImg.src = imgSrc;
+            elements.previewImg.style.display = 'block';
+            
+            // Show filename
+            if (elements.imageName) {
+                elements.imageName.textContent = filename;
+                elements.imageName.style.display = 'block';
+            }
+        };
+        img.onerror = () => {
+            clearTimeout(spinnerTimeout);
+            elements.loadingSpinner.style.display = 'none';
+            showMessage(MESSAGES.ERROR);
+            
+            // Hide filename on error
+            if (elements.imageName) {
+                elements.imageName.style.display = 'none';
+            }
+        };
+        img.src = imgSrc;
+    };
+
+    const updateTagCountBadges = () => {
+        document.querySelectorAll('[data-tag-count]').forEach(badge => {
+            const imageName = badge.getAttribute('data-tag-count');
+            const count = editor.countPageMarkerOccurrences(imageName);
+            if (count > 1) {
+                badge.textContent = `×${count}`;
+                badge.style.display = 'inline';
+                badge.title = `Cette balise apparaît ${count} fois dans le document`;
+            } else {
+                badge.style.display = 'none';
+            }
+        });
+    };
+
+    const refreshButtonStates = () => {
+        deactivateActiveButton();
+        
+        // Update all buttons to their proper state
+        document.querySelectorAll('.editor [data-tag]').forEach(button => {
+            const imageName = button.getAttribute('data-tag');
+            const isInserted = editor.isPageMarkerInserted(imageName);
+            
+            const state = isInserted ? BUTTON_STATES.INSERTED : BUTTON_STATES.INACTIVE;
+            setButtonState(button, state);
+            
+            button.querySelector('span').textContent = imageName;
+            
+            if (isInserted) {
+                button.setAttribute('data-inserted', 'true');
+            } else {
+                button.removeAttribute('data-inserted');
+            }
+        });
+        updateTagCountBadges();
+        updatePaginationColors();
+    };
+
+    // Event handlers
+    elements.toggleBtn.addEventListener('click', () => {
+        const isReadOnly = editor.toggleReadOnly();
+        elements.toggleBtn.textContent = isReadOnly ? 'Activer le mode édition' : 'Activer le mode lecture seule';
+        elements.toggleBtn.classList.toggle('btn-warning');
+        elements.toggleBtn.classList.toggle('btn-info');
+
+        document.querySelectorAll('[data-enable-when-readonly]').forEach(btn => {
+            btn.disabled = !isReadOnly;
+        });
+
+        if (isReadOnly) {
+            refreshButtonStates();
+        } else {
+            deactivateActiveButton();
+        }
     });
 
-    const data = await response.json();
+    elements.toggleTagsBtn.addEventListener('click', () => {
+        const tagsHidden = editor.toggleTagVisibility();
+        elements.toggleTagsBtn.textContent = tagsHidden ? 'Afficher les balises' : 'Masquer les balises';
+        elements.toggleTagsBtn.classList.toggle('btn-secondary');
+        elements.toggleTagsBtn.classList.toggle('btn-info');
+    });
 
-    if (response.status === 403) {
-      alert(data.error || 'Modification non autorisée : cette comparaison est publiée.');
-    } else if (response.ok) {
-      alert(data.message || 'Fichier sauvegardé avec succès !');
-    } else {
-      alert(data.error || 'Une erreur est survenue lors de la sauvegarde.');
+    // Insert buttons
+    document.querySelectorAll('.editor [data-tag]').forEach(button => {
+        const imgSrc = button.getAttribute('data-img-src');
+        const imageName = button.getAttribute('data-tag');
+
+        button.addEventListener('click', () => {
+            const isInserted = editor.isPageMarkerInserted(imageName);
+            
+            if (activeButton === button) {
+                if (isDeleteMode) {
+                    editor.removePageMarker(imageName);
+                    refreshButtonStates();
+                } else {
+                    deactivateActiveButton();
+                }
+                return;
+            }
+            
+            if (activeButton) {
+                deactivateActiveButton();
+            }
+            
+            activeButton = button;
+            loadImage(imgSrc);
+            
+            if (isInserted) {
+                isDeleteMode = true;
+                setButtonState(button, BUTTON_STATES.ACTIVE_DELETE);
+                button.querySelector('span').textContent = '🗑️';
+                editor.scrollToPageMarker(imageName);
+                elements.editorContainer.style.cursor = 'default';
+            } else {
+                isDeleteMode = false;
+                setButtonState(button, BUTTON_STATES.ACTIVE_INSERT);
+                button.querySelector('span').textContent = imageName;
+                elements.editorContainer.style.cursor = 'crosshair';
+            }
+        });
+
+        // Hover to show image preview
+        button.addEventListener('mouseenter', () => {
+            if (imgSrc) {
+                loadImage(imgSrc);
+            }
+        });
+
+        // On mouse leave, restore active button image if there is one
+        button.addEventListener('mouseleave', () => {
+            // If leaving the active button in delete mode, deactivate it
+            if (activeButton === button && isDeleteMode) {
+                deactivateActiveButton();
+            } else if (activeButton && activeButton !== button) {
+                const activeImgSrc = activeButton.getAttribute('data-img-src');
+                loadImage(activeImgSrc);
+            }
+        });
+    });
+
+    // Editor click to insert tag
+    if (elements.editorContainer) {
+        elements.editorContainer.addEventListener('click', () => {
+            // Only insert if we have an active button in insert mode
+            if (activeButton && !isDeleteMode) {
+                const imageName = activeButton.getAttribute('data-tag');
+                const pageNumber = '001'; // Hard coded for now
+                editor.insertPageMarker(imageName, pageNumber);
+                refreshButtonStates();
+            }
+        });
     }
-  });
-};
+
+    elements.saveBtn.addEventListener('click', async () => {
+        const updatedXml = view.state.doc.toString();
+        const response = await fetch(`/comparison/${comparisonId}/editor?type=${fileType}`, {
+            method: 'PUT',
+            headers: {
+            'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content'),
+            'Content-Type': 'application/xml',
+            'Accept': 'application/json'
+            },
+            body: updatedXml
+        });
+
+        const data = await response.json();
+
+        if (response.status === 403) {
+            alert(data.error || 'Modification non autorisée : cette comparaison est publiée.');
+        } else if (response.ok) {
+            alert(data.message || 'Fichier sauvegardé avec succès !');
+        } else {
+            alert(data.error || 'Une erreur est survenue lors de la sauvegarde.');
+        }
+    });
+
+    initPagination();
+    refreshButtonStates();
+});
