+++ laravel/resources/views/components/main/editor.blade.php
@@ -107,363 +107,12 @@
 
 @push('scripts')
     <script>
-        document.addEventListener('DOMContentLoaded', () => {
-            const xmlContent = @json($xmlContent);
-            const comparisonId = {{ $comparison->id }};
-            const fileType = '{{ $isSource ? 'source' : 'target' }}';
-            const canEdit = {{ $canEdit ? 'true' : 'false' }};
-
-            // DOM Elements
-            const elements = {
-                toggleBtn: document.getElementById('toggle-readonly'),
-                toggleTagsBtn: document.getElementById('toggle-tags'),
-                previewImg: document.getElementById('facsimile-preview'),
-                noPreviewText: document.getElementById('no-preview'),
-                loadingSpinner: document.getElementById('loading-spinner'),
-                editorContainer: document.getElementById('editor-container'),
-                imageName: document.getElementById('image-name'),
-                pagination: document.getElementById('pagination'),
-            };
-
-            // Constants
-            const MESSAGES = {
-                DEFAULT: 'Cliquez sur un bouton "Insérer..." pour activer l\'insertion',
-                ERROR: 'Erreur lors du chargement de l\'image'
-            };
-
-            const BUTTON_STATES = {
-                INACTIVE: { add: 'btn-primary', remove: ['btn-warning', 'btn-success', 'btn-danger'] },
-                ACTIVE_INSERT: { add: 'btn-warning', remove: ['btn-primary', 'btn-success', 'btn-danger'] },
-                INSERTED: { add: 'btn-success', remove: ['btn-primary', 'btn-warning', 'btn-danger'] },
-                ACTIVE_DELETE: { add: 'btn-danger', remove: ['btn-primary', 'btn-warning', 'btn-success'] }
-            };
-
-            // State - only one active button at a time, either in insert or delete mode
-            let activeButton = null;
-            let isDeleteMode = false; // true = delete mode, false = insert mode
-            let lastImageSrc = null; // Track the last image displayed
-            let currentPage = 1;
-            const itemsPerPage = 3;
-
-            window.initEditor(xmlContent, comparisonId, fileType);
-            if (!canEdit && window.editor) {
-                window.editor.setReadOnly(true);
-            }
-
-            // Pagination functions
-            const checkIfPageFullyInserted = (pageNumber) => {
-                const pageButtons = document.querySelectorAll(`.button-item[data-page="${pageNumber}"] button[data-tag]`);
-                if (pageButtons.length === 0) return false;
-                
-                let allInserted = true;
-                pageButtons.forEach(button => {
-                    const imageName = button.getAttribute('data-tag');
-                    if (window.editor && !window.editor.isPageMarkerInserted(imageName)) {
-                        allInserted = false;
-                    }
-                });
-                
-                return allInserted;
-            };
-
-            const updatePaginationColors = () => {
-                document.querySelectorAll('#pagination .page-item').forEach((item, index) => {
-                    const pageNumber = index + 1;
-                    const link = item.querySelector('.page-link');
-                    const isFullyInserted = checkIfPageFullyInserted(pageNumber);
-                    
-                    if (isFullyInserted) {
-                        link.classList.add('page-fully-inserted');
-                    } else {
-                        link.classList.remove('page-fully-inserted');
-                    }
-                });
-            };
-
-            const initPagination = () => {
-                const allButtons = document.querySelectorAll('.button-item');
-                const totalPages = Math.ceil(allButtons.length / itemsPerPage);
-                
-                allButtons.forEach((item, index) => {
-                    const itemPage = Math.ceil((index + 1) / itemsPerPage);
-                    item.setAttribute('data-page', itemPage);
-                });
-                
-                if (totalPages <= 1) {
-                    elements.pagination.parentElement.style.display = 'none';
-                    showPage(1);
-                    return;
-                }
-                
-                elements.pagination.innerHTML = '';
-                
-                for (let i = 1; i <= totalPages; i++) {
-                    const li = document.createElement('li');
-                    li.className = 'page-item';
-                    
-                    const link = document.createElement('a');
-                    link.className = 'page-link shadow-none';
-                    link.href = '#';
-                    link.textContent = i;
-                    link.addEventListener('click', (e) => {
-                        e.preventDefault();
-                        showPage(i);
-                    });
-                    
-                    li.appendChild(link);
-                    elements.pagination.appendChild(li);
-                }
-                
-                showPage(1);
-            };
-
-            const showPage = (page) => {
-                currentPage = page;
-                
-                const allButtons = document.querySelectorAll('.button-item');
-                allButtons.forEach((item) => {
-                    const itemPage = parseInt(item.getAttribute('data-page'));
-                    item.style.display = itemPage === page ? '' : 'none';
-                });
-                
-                // Update pagination active state
-                document.querySelectorAll('#pagination .page-item').forEach((item, index) => {
-                    if (index + 1 === page) {
-                        item.classList.add('active');
-                    } else {
-                        item.classList.remove('active');
-                    }
-                });
-            };
-
-            // Utility functions
-            const setButtonState = (button, state) => {
-                if (Array.isArray(state.remove)) {
-                    button.classList.remove(...state.remove);
-                } else {
-                    button.classList.remove(state.remove);
-                }
-                button.classList.add(state.add);
-            };
-
-            const showMessage = (message) => {
-                if (elements.noPreviewText) {
-                    elements.noPreviewText.textContent = message;
-                    elements.noPreviewText.style.display = 'block';
-                }
-            };
-            showMessage(MESSAGES.DEFAULT);
-
-            const hidePreview = () => {
-                // Only hide if no image has been displayed yet
-                if (!lastImageSrc) {
-                    elements.previewImg.style.display = 'none';
-                    if (elements.imageName) elements.imageName.style.display = 'none';
-                    showMessage(MESSAGES.DEFAULT);
-                }
-            };
-
-            const deactivateActiveButton = () => {
-                if (activeButton) {
-                    const imageName = activeButton.getAttribute('data-tag');
-                    const isInserted = window.editor && window.editor.isPageMarkerInserted(imageName);
-                    
-                    // Reset to appropriate state
-                    if (isInserted) {
-                        setButtonState(activeButton, BUTTON_STATES.INSERTED);
-                    } else {
-                        setButtonState(activeButton, BUTTON_STATES.INACTIVE);
-                    }
-                    activeButton.querySelector('span').textContent = imageName;
-                    
-                    activeButton = null;
-                    isDeleteMode = false;
-                    elements.editorContainer.style.cursor = 'default';
-                }
-            };
-
-            const loadImage = (imgSrc) => {
-                if (!imgSrc) return;
-                
-                lastImageSrc = imgSrc;
-                if (elements.noPreviewText) elements.noPreviewText.style.display = 'none';
-                
-                // Extract filename from URL
-                const filename = imgSrc.split('/').pop();
-                
-                // Set a timeout to show spinner only if loading takes more than 500ms
-                let spinnerTimeout = setTimeout(() => {
-                    elements.previewImg.style.display = 'none';
-                    elements.loadingSpinner.style.display = 'block';
-                }, 500);
-                
-                const img = new Image();
-                img.onload = () => {
-                    clearTimeout(spinnerTimeout);
-                    elements.loadingSpinner.style.display = 'none';
-                    elements.previewImg.src = imgSrc;
-                    elements.previewImg.style.display = 'block';
-                    
-                    // Show filename
-                    if (elements.imageName) {
-                        elements.imageName.textContent = filename;
-                        elements.imageName.style.display = 'block';
-                    }
-                };
-                img.onerror = () => {
-                    clearTimeout(spinnerTimeout);
-                    elements.loadingSpinner.style.display = 'none';
-                    showMessage(MESSAGES.ERROR);
-                    
-                    // Hide filename on error
-                    if (elements.imageName) {
-                        elements.imageName.style.display = 'none';
-                    }
-                };
-                img.src = imgSrc;
-            };
-
-            const updateTagCountBadges = () => {
-                document.querySelectorAll('[data-tag-count]').forEach(badge => {
-                    const imageName = badge.getAttribute('data-tag-count');
-                    if (window.editor) {
-                        const count = window.editor.countPageMarkerOccurrences(imageName);
-                        if (count > 1) {
-                            badge.textContent = `×${count}`;
-                            badge.style.display = 'inline';
-                            badge.title = `Cette balise apparaît ${count} fois dans le document`;
-                        } else {
-                            badge.style.display = 'none';
-                        }
-                    }
-                });
-            };
-
-            const refreshButtonStates = () => {
-                deactivateActiveButton();
-                
-                // Update all buttons to their proper state
-                document.querySelectorAll('.editor [data-tag]').forEach(button => {
-                    const imageName = button.getAttribute('data-tag');
-                    const isInserted = window.editor.isPageMarkerInserted(imageName);
-                    
-                    const state = isInserted ? BUTTON_STATES.INSERTED : BUTTON_STATES.INACTIVE;
-                    setButtonState(button, state);
-                    
-                    button.querySelector('span').textContent = imageName;
-                    
-                    if (isInserted) {
-                        button.setAttribute('data-inserted', 'true');
-                    } else {
-                        button.removeAttribute('data-inserted');
-                    }
-                });
-                updateTagCountBadges();
-                updatePaginationColors();
-            };
-
-            // Event handlers
-            elements.toggleBtn.addEventListener('click', () => {
-                const isReadOnly = window.editor.toggleReadOnly();
-                elements.toggleBtn.textContent = isReadOnly ? 'Activer le mode édition' : 'Activer le mode lecture seule';
-                elements.toggleBtn.classList.toggle('btn-warning');
-                elements.toggleBtn.classList.toggle('btn-info');
-
-                document.querySelectorAll('[data-enable-when-readonly]').forEach(btn => {
-                    btn.disabled = !isReadOnly;
-                });
-
-                if (isReadOnly) {
-                    refreshButtonStates();
-                } else {
-                    deactivateActiveButton();
-                }
-            });
-
-            elements.toggleTagsBtn.addEventListener('click', () => {
-                const tagsHidden = window.editor.toggleTagVisibility();
-                elements.toggleTagsBtn.textContent = tagsHidden ? 'Afficher les balises' : 'Masquer les balises';
-                elements.toggleTagsBtn.classList.toggle('btn-secondary');
-                elements.toggleTagsBtn.classList.toggle('btn-info');
-            });
-
-            // Insert buttons
-            document.querySelectorAll('.editor [data-tag]').forEach(button => {
-                const imgSrc = button.getAttribute('data-img-src');
-                const imageName = button.getAttribute('data-tag');
-
-                button.addEventListener('click', () => {
-                    if (!window.editor) {
-                        return;
-                    }
-                    
-                    const isInserted = window.editor.isPageMarkerInserted(imageName);
-                    
-                    if (activeButton === button) {
-                        if (isDeleteMode) {
-                            window.editor.removePageMarker(imageName);
-                            refreshButtonStates();
-                        } else {
-                            deactivateActiveButton();
-                        }
-                        return;
-                    }
-                    
-                    if (activeButton) {
-                        deactivateActiveButton();
-                    }
-                    
-                    activeButton = button;
-                    loadImage(imgSrc);
-                    
-                    if (isInserted) {
-                        isDeleteMode = true;
-                        setButtonState(button, BUTTON_STATES.ACTIVE_DELETE);
-                        button.querySelector('span').textContent = '🗑️';
-                        window.editor.scrollToPageMarker(imageName);
-                        elements.editorContainer.style.cursor = 'default';
-                    } else {
-                        isDeleteMode = false;
-                        setButtonState(button, BUTTON_STATES.ACTIVE_INSERT);
-                        button.querySelector('span').textContent = imageName;
-                        elements.editorContainer.style.cursor = 'crosshair';
-                    }
-                });
-
-                // Hover to show image preview
-                button.addEventListener('mouseenter', () => {
-                    if (imgSrc) {
-                        loadImage(imgSrc);
-                    }
-                });
-
-                // On mouse leave, restore active button image if there is one
-                button.addEventListener('mouseleave', () => {
-                    // If leaving the active button in delete mode, deactivate it
-                    if (activeButton === button && isDeleteMode) {
-                        deactivateActiveButton();
-                    } else if (activeButton && activeButton !== button) {
-                        const activeImgSrc = activeButton.getAttribute('data-img-src');
-                        loadImage(activeImgSrc);
-                    }
-                });
-            });
-
-            // Editor click to insert tag
-            if (elements.editorContainer) {
-                elements.editorContainer.addEventListener('click', () => {
-                    // Only insert if we have an active button in insert mode
-                    if (activeButton && !isDeleteMode && window.editor) {
-                        const imageName = activeButton.getAttribute('data-tag');
-                        const pageNumber = '001'; // Hard coded for now
-                        window.editor.insertPageMarker(imageName, pageNumber);
-                        refreshButtonStates();
-                    }
-                });
-            }
-
-            initPagination();
-            refreshButtonStates();
-        });
+      window.editorParams = {
+        xmlContent: @json($xmlContent),
+        comparisonId: {{ $comparison->id }},
+        fileType: '{{ $isSource ? 'source' : 'target' }}',
+        canEdit: {{ $canEdit ? 'true' : 'false' }},
+      };
     </script>
+    @vite('resources/js/editor.js')
 @endpush
