+++ laravel/vite.config.js
@@ -4,7 +4,11 @@ import laravel from 'laravel-vite-plugin'
 export default defineConfig({
   plugins: [
     laravel({
-      input: ['resources/js/app.js', 'resources/css/app.css'],
+      input: [
+        'resources/js/app.js',
+        'resources/css/app.css',
+        'resources/js/editor.js',
+      ],
       refresh: true,          // harmless in prod builds; useful if you run `vite` in dev
       buildDirectory: 'build' // outputs to public/build (default), explicit for clarity
     }),
